requirements.txt
argcomplete==3.0.8
asgiref==3.7.1
backports.zoneinfo==0.2.1
boto3==1.26.141
botocore==1.29.141
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Django==4.2.1
djangorestframework==3.14.0
durationpy==0.5
flake8==6.0.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kappa==0.6.0
MarkupSafe==2.1.2
mccabe==0.7.0
nltk==3.8.1
placebo==0.9.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.31.0
ruby==0.0.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
typing-extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
zappa==0.57.0
